+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - 0.1.2
+    - Fixed Grid Search best hyper-parameters method
     - Fixed K Means average loss calculation
     - Fixed bootstrap estimators tiny bootstrap sets